v0.3- Navbar iFrame
Included the navbar in an iFrame on the pages that I am responsible for to improve consistency and make implementing changes easier. Functionality for adding parameters for search to search.html and settings.html added, but settings.html still needs additional attention to facilitate sending information from the form using POST to the server.